break up inventory search component into outer + result components
